requirements.txt
Django==4.2
psycopg2-binary==2.9.9
django-environ==0.11.2
pillow==10.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
factory-boy==3.3.0
Faker==24.2.0